moved bcolors to helpers
