+++ stag/check_create_db_input_files.py
@@ -8,22 +8,11 @@ import subprocess
 import tempfile
 import shlex
 
-from stag.helpers import is_tool, linearise_fasta
+from stag.helpers import is_tool, linearise_fasta, bcolors
 
 # there are two input files to check:
 # - the taxonomy file
 # - the fasta file
-
-class bcolors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
 
 # ------------------------------------------------------------------------------
 # 1. check taxonomy
