+++ stag/helpers.py
@@ -3,6 +3,16 @@ import sys
 import shlex
 import subprocess
 
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
 # colors for the shell ---------------------------------------------------------
 class bco:
 	ResetAll = "\033[0m"
